# UCT/CUDA_IPC: Fix multiple calls to `cuIpcCloseMemHandle`