## What
Fix multiple calls to `cuIpcCloseMemHandle` on same handle.

## Why ?
Multiple calls to `cuIpcCloseMemHandle` occur when transferring memory buffers larger than 1GB, as such buffers will be copied in 1GB blocks and each will attempt to unmap the memory handle that is still at use.

## How ?
Add a callback function to the `uct_cuda_ipc_key` struct that only gets called upon release of the object.