@pentschev 2019-11-11
@bureddy this fixes the multiple calls to `cuIpcCloseMemHandle` that we discussed in https://github.com/openucx/ucx/issues/4410. After spending time understanding the code, I figured that this only happens when the memory buffer being transferred is > 1GB, for which the copy will be dispatched in multiple 1GB chunks and each will attempt to unmap the handle.

cc @quasiben