@mellanox-github 2019-11-12
<details><summary>Mellanox CI: <b>PASSED</b#  > on 25 workers (click for details)</summary>

**Note:** the logs will be deleted after 19-Nov-2019

Agent/Stage | Status
-- | --
[_main](https://gist.githubusercontent.com/mellanox-github/084f89036625e263e7c0a35fbeada5cc/raw/f645da5e7be3529b8b4585b150fdec2bc3808e1b/_main) | :heavy_check_mark: SUCCESS
[hpc-arm-cavium-jenkins_W0](https://gist.githubusercontent.com/mellanox-github/084f89036625e263e7c0a35fbeada5cc/raw/9d80e45c5710f7a24d351d8b82c5b6d5e476c84a/hpc-arm-cavium-jenkins_W0) | :heavy_check_mark: SUCCESS
[hpc-arm-cavium-jenkins_W1](https://gist.githubusercontent.com/mellanox-github/084f89036625e263e7c0a35fbeada5cc/raw/05b34f6811ad5ff3c5ec5ad214c430dd7c3e9fa5/hpc-arm-cavium-jenkins_W1) | :heavy_check_mark: SUCCESS
[hpc-arm-cavium-jenkins_W2](https://gist.githubusercontent.com/mellanox-github/084f89036625e263e7c0a35fbeada5cc/raw/f57d9d9d2ac7663aa634021bacbfbd3deee279b3/hpc-arm-cavium-jenkins_W2) | :heavy_check_mark: SUCCESS
[hpc-arm-cavium-jenkins_W3](https://gist.githubusercontent.com/mellanox-github/084f89036625e263e7c0a35fbeada5cc/raw/b9300aa8a55a1565406d3ac28578e2094bed9ed9/hpc-arm-cavium-jenkins_W3) | :heavy_check_mark: SUCCESS
[hpc-arm-hwi-jenkins_W0](https://gist.githubusercontent.com/mellanox-github/084f89036625e263e7c0a35fbeada5cc/raw/7a7c38531a44165fe3dede419129e00b5cfda272/hpc-arm-hwi-jenkins_W0) | :heavy_check_mark: SUCCESS
[hpc-arm-hwi-jenkins_W1](https://gist.githubusercontent.com/mellanox-github/084f89036625e263e7c0a35fbeada5cc/raw/98e13c03dede3bbf5cdb1fbd6fc2ec4612f0e94f/hpc-arm-hwi-jenkins_W1) | :heavy_check_mark: SUCCESS
[hpc-arm-hwi-jenkins_W2](https://gist.githubusercontent.com/mellanox-github/084f89036625e263e7c0a35fbeada5cc/raw/c8d742c29cff77336124e2df7a587d13975422d0/hpc-arm-hwi-jenkins_W2) | :heavy_check_mark: SUCCESS
[hpc-arm-hwi-jenkins_W3](https://gist.githubusercontent.com/mellanox-github/084f89036625e263e7c0a35fbeada5cc/raw/9f4240e32508adefb87bfd7ef25fe071e63efd09/hpc-arm-hwi-jenkins_W3) | :heavy_check_mark: SUCCESS
[hpc-test-node-gpu_W0](https://gist.githubusercontent.com/mellanox-github/084f89036625e263e7c0a35fbeada5cc/raw/476278b7572fa49f5e78443eaafdbe1be581669e/hpc-test-node-gpu_W0) | :heavy_check_mark: SUCCESS
[hpc-test-node-gpu_W1](https://gist.githubusercontent.com/mellanox-github/084f89036625e263e7c0a35fbeada5cc/raw/ea413c6ad26897230fedf87d99a6e9e11f384869/hpc-test-node-gpu_W1) | :heavy_check_mark: SUCCESS
[hpc-test-node-gpu_W2](https://gist.githubusercontent.com/mellanox-github/084f89036625e263e7c0a35fbeada5cc/raw/71608fc343dccd4f7c5fac4b5bc5e818e2ec4e20/hpc-test-node-gpu_W2) | :heavy_check_mark: SUCCESS
[hpc-test-node-gpu_W3](https://gist.githubusercontent.com/mellanox-github/084f89036625e263e7c0a35fbeada5cc/raw/de1bf29341c78f95ca0eccdcb85f97650391e101/hpc-test-node-gpu_W3) | :heavy_check_mark: SUCCESS
[hpc-test-node-legacy_W0](https://gist.githubusercontent.com/mellanox-github/084f89036625e263e7c0a35fbeada5cc/raw/b80dbf08698b4b472b3d1fa808f973c93cceeb7a/hpc-test-node-legacy_W0) | :heavy_check_mark: SUCCESS
[hpc-test-node-legacy_W1](https://gist.githubusercontent.com/mellanox-github/084f89036625e263e7c0a35fbeada5cc/raw/36fb0914b9fcc5662aaf8524d4630c1fa90f314b/hpc-test-node-legacy_W1) | :heavy_check_mark: SUCCESS
[hpc-test-node-legacy_W2](https://gist.githubusercontent.com/mellanox-github/084f89036625e263e7c0a35fbeada5cc/raw/5559de64d72dcab0c4ea47aeec7b1b3998d08912/hpc-test-node-legacy_W2) | :heavy_check_mark: SUCCESS
[hpc-test-node-legacy_W3](https://gist.githubusercontent.com/mellanox-github/084f89036625e263e7c0a35fbeada5cc/raw/47ac07b20b4ba4a727f7f750302f91b74e667422/hpc-test-node-legacy_W3) | :heavy_check_mark: SUCCESS
[hpc-test-node-new_W0](https://gist.githubusercontent.com/mellanox-github/084f89036625e263e7c0a35fbeada5cc/raw/5f79671593bf13e19ed5ea3f52fa131b718b2a0d/hpc-test-node-new_W0) | :heavy_check_mark: SUCCESS
[hpc-test-node-new_W1](https://gist.githubusercontent.com/mellanox-github/084f89036625e263e7c0a35fbeada5cc/raw/dfdb1706c8ebf188d70688f4ab422c304eb6f32b/hpc-test-node-new_W1) | :heavy_check_mark: SUCCESS
[hpc-test-node-new_W2](https://gist.githubusercontent.com/mellanox-github/084f89036625e263e7c0a35fbeada5cc/raw/fe7b40d3f66af5d28e924e2e42550f2589417212/hpc-test-node-new_W2) | :heavy_check_mark: SUCCESS
[hpc-test-node-new_W3](https://gist.githubusercontent.com/mellanox-github/084f89036625e263e7c0a35fbeada5cc/raw/b265c36c1a31a2f0c7599ed2d6d078f6f6708b80/hpc-test-node-new_W3) | :heavy_check_mark: SUCCESS
[r-vmb-ppc-jenkins_W0](https://gist.githubusercontent.com/mellanox-github/084f89036625e263e7c0a35fbeada5cc/raw/641372e2cb1cc0345b380842fb26dd438681454d/r-vmb-ppc-jenkins_W0) | :heavy_check_mark: SUCCESS
[r-vmb-ppc-jenkins_W1](https://gist.githubusercontent.com/mellanox-github/084f89036625e263e7c0a35fbeada5cc/raw/951c638bbd9a6637b80b096a9be1726fce77d625/r-vmb-ppc-jenkins_W1) | :heavy_check_mark: SUCCESS
[r-vmb-ppc-jenkins_W2](https://gist.githubusercontent.com/mellanox-github/084f89036625e263e7c0a35fbeada5cc/raw/8a2518a88671342345510004fcfa9e5c2f748b5f/r-vmb-ppc-jenkins_W2) | :heavy_check_mark: SUCCESS
[r-vmb-ppc-jenkins_W3](https://gist.githubusercontent.com/mellanox-github/084f89036625e263e7c0a35fbeada5cc/raw/84d44ad02275556602945b33d91803263f82d088/r-vmb-ppc-jenkins_W3) | :heavy_check_mark: SUCCESS