@pentschev 2019-11-12
All changes requested have ben addressed now.