@brminich 2019-11-12
looks like the comments are not addressed