@pentschev 2019-11-12
> looks like the comments are not addressed

@brminich could you point out which one(s)? I just went through all the comments once again and I couldn't find any that have not been addressed.