@brminich 2019-11-12
The latest push contains all the fixes. Now I can see them.
thanks!