@pentschev 2019-11-13
> @pentschev Can you also post if the following work and yields expected performance?

@Akshay-Venkatesh sure, here are the results I get.

Before this PR with cache:

```
[1573633946.874246] [dgx11:29267:0]       perftest.c:1380 UCX  WARN  CPU affinity is not set (bound to 80 cpus). Performance may be impacted.
+--------------+-----------------------------+---------------------+-----------------------+
|              |       latency (usec)        |   bandwidth (MB/s)  |  message rate (msg/s) |
+--------------+---------+---------+---------+----------+----------+-----------+-----------+
| # iterations | typical | average | overall |  average |  overall |   average |   overall |
+--------------+---------+---------+---------+----------+----------+-----------+-----------+
           100     0.000    55.380    55.380   17220.60   17220.60       18057       18057
```

After this PR with cache:

```
[1573634633.822003] [dgx11:59875:0]       perftest.c:1380 UCX  WARN  CPU affinity is not set (bound to 80 cpus). Performance may be impacted.
+--------------+-----------------------------+---------------------+-----------------------+
|              |       latency (usec)        |   bandwidth (MB/s)  |  message rate (msg/s) |
+--------------+---------+---------+---------+----------+----------+-----------+-----------+
| # iterations | typical | average | overall |  average |  overall |   average |   overall |
+--------------+---------+---------+---------+----------+----------+-----------+-----------+
           100     0.000    55.070    55.070   17317.52   17317.52       18159       18159
```

While trying cache disabled, I noticed that numbers vary widely with only 100 iterations, but we get more stable overall results starting at about 1-2k iterations, so the results I get for 2k are below:

Before this PR without cache:
```
[1573635432.345461] [dgx11:72299:0]       perftest.c:1380 UCX  WARN  CPU affinity is not set (bound to 80 cpus). Performance may be impacted.
+--------------+-----------------------------+---------------------+-----------------------+
|              |       latency (usec)        |   bandwidth (MB/s)  |  message rate (msg/s) |
+--------------+---------+---------+---------+----------+----------+-----------+-----------+
| # iterations | typical | average | overall |  average |  overall |   average |   overall |
+--------------+---------+---------+---------+----------+----------+-----------+-----------+
           451     0.000  2223.807  2223.807     428.85     428.85         450         450
           710     0.000  3870.437  2824.479     246.40     337.65         258         354
          1247   981.280  1867.409  2412.333     510.69     395.33         536         415
          1456   981.280  4806.293  2755.971     198.42     346.04         208         363
          2000   946.808  1726.344  2475.913     552.42     385.18         579         404
```

After this PR without cache:
```
[1573635461.461535] [dgx11:74630:0]       perftest.c:1380 UCX  WARN  CPU affinity is not set (bound to 80 cpus). Performance may be impacted.
+--------------+-----------------------------+---------------------+-----------------------+
|              |       latency (usec)        |   bandwidth (MB/s)  |  message rate (msg/s) |
+--------------+---------+---------+---------+----------+----------+-----------+-----------+
| # iterations | typical | average | overall |  average |  overall |   average |   overall |
+--------------+---------+---------+---------+----------+----------+-----------+-----------+
           573     0.000  1752.859  1752.859     544.07     544.07         570         570
           791     0.000  5060.408  2664.421     188.46     357.93         198         375
          1253   977.491  2169.857  2482.068     439.51     384.23         461         403
          1592   977.491  2979.401  2587.970     320.09     368.50         336         386
          2000   952.565  2213.821  2511.644     430.78     379.70         452         398
```

Performance with this PR seems very much the same to me.